Move ModelFactory to own file
This fixes a circular import issue that happens when we try to import
ContextImplications from ggrc_basic_permissions.models.factories.
